+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### 0.2.0 (Next)
 
+* Added support for operators with `operator [name]` - [@dblock](https://github.com/dblock).
+* Added support for custom commands with `command [name]` - [@dblock](https://github.com/dblock).
 * Your contribution here.
 
 ### 0.1.0 (6/2/2015)